requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.22
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0

# Authentication & Security
python-jose>=3.3.0
passlib==1.7.4
bcrypt==3.2.2

# Utilities
tenacity>=8.2.3
pyyaml>=6.0.1
email-validator>=2.0.0
Pillow>=10.0.0  # For image processing
python-magic>=0.4.27  # For file type detection

# Testing
pytest>=7.4.2
httpx>=0.25.0
pytest-asyncio>=0.21.1

# Development
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.1